refactor: Place state in Rolling
